% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{calculate_features}
\alias{calculate_features}
\title{Calcula les dades de la malla}
\usage{
calculate_features(grid_sf, mde, slope, aspect, coastline)
}
\arguments{
\item{grid_sf}{Objecte sf de punts, obtingut amb \link{create_grid_sf}` o data.frama amb coordenades com l'obtingut amb \link[meteocat]{create_grid_tidy}}

\item{mde}{Arxiu ràster que conté un model digital d'elevacions. Es pot obtenir amb \link[meteocat]{get_mde}.}

\item{slope}{Arxiu raster de pendents. Es pot obtenir amb \link{get_terrain}.}

\item{aspect}{Arxiu raster d'orientacions. Es pot obtenir amb \link{get_terrain}.}

\item{coastline}{Objecte sf que representa la línia de costa.}
}
\value{
Una tidytable amb les mateixes files que \code{grid_sf} i amb columnes corresponents a les dades utilitzades per \link{interpolate}
}
\description{
Calcula les dades de la malla
}
