% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid_sf}
\alias{create_grid_sf}
\title{Crea una malla de punts per la interpolació}
\usage{
create_grid_sf(area_estudi)
}
\arguments{
\item{area_estudi}{Arxiu raster amb l'àrea d'estudi}
}
\value{
Un objecte sf amb un punt per cada píxel del raster \code{area_estudi}
}
\description{
Crea una malla de punts per la interpolació
}
