% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid_tidy}
\alias{create_grid_tidy}
\title{Crea una tidytable de coordenades a partir d'una malla}
\usage{
create_grid_tidy(grid_sf)
}
\arguments{
\item{grid_sf}{Objecte sf de punts, obtingut amb \link[meteocat]{create_grid_sf}`}
}
\value{
Una tidytable amb coordenades i un identificador únic.
}
\description{
Crea una tidytable de coordenades a partir d'una malla
}
