% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{distance_to_sea}
\alias{distance_to_sea}
\title{Calcula la distància al mar de cada punt de la malla}
\usage{
distance_to_sea(grid_sf, coastline)
}
\arguments{
\item{grid_sf}{Objecte sf de punts, obtingut amb \link[meteocat]{create_grid_sf}`}

\item{coastline}{Objecte sf representant la línia de costa}
}
\value{
Un vector numèric amb les distàncies
}
\description{
Calcula la distància al mar de cada punt de la malla
}
