% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_code_variables.R
\name{download_code_variables}
\alias{download_code_variables}
\title{Descarrega els codis de les variables}
\usage{
download_code_variables(api_key = get_api_key())
}
\arguments{
\item{api_key}{La clau API del meteocat. Si es guarda a .Renviron com a 'meteocat_key' s'estira automàticament amb \code{get_api_key()}}
}
\value{
Una tidytable amb codi, nom, unitat, acrònim, tipus i decimals.
}
\description{
Descarrega els codis de les variables
}
