% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_stations.R
\name{download_stations}
\alias{download_stations}
\title{Descarrega dades de les estacions de la XEMA}
\usage{
download_stations(api_key = get_api_key())
}
\arguments{
\item{api_key}{La clau API del meteocat. Si es guarda a .Renviron com a 'meteocat_key' s'estira automàticament amb \code{get_api_key()}}
}
\value{
Un objecte sf amb les estacions meteorològiques de la XEMA
}
\description{
Descarrega dades de les estacions de la XEMA
}
