% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_var.R
\name{download_var}
\alias{download_var}
\title{Descarrega la mitjana d'una sèrie multianual}
\usage{
download_var(
  variable,
  years,
  variables = download_code_variables(),
  api_key = get_api_key()
)
}
\arguments{
\item{variable}{Acrònim de la variable obtinguda de \code{download_code_variables()}}

\item{years}{Un vector amb tots els anys que volem incloure a la sèrie temporal}

\item{variables}{La taula de variables que s'obté amb \code{download_code_variables()}}

\item{api_key}{La clau API del meteocat. Si es guarda a .Renviron com a 'meteocat_key' s'estira automàticament amb \code{get_api_key()}}
}
\value{
Una tidytable amb una columna per mes i una fila per estació.
}
\description{
Descarrega la mitjana d'una sèrie multianual
}
\details{
La funció retorna el valor mitjà dels anys inclosos a la sèrie pels quals hi ha dades. Si una estació no té dades per cap dels anys, s'elimina de la taula.
}
