% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_var_year.R
\name{download_var_year}
\alias{download_var_year}
\title{Descarrega una variable per un any concret}
\usage{
download_var_year(code_variable, year, api_key = get_api_key())
}
\arguments{
\item{code_variable}{Codi de la variable obtingut com \code{download_code_variables()}}

\item{year}{Any que es vol descarregar}

\item{api_key}{La clau API del meteocat. Si es guarda a .Renviron com a 'meteocat_key' s'estira automàticament amb \code{get_api_key()}}
}
\value{
Una tidytable amb una columna per mes i una fila per estació.
}
\description{
Descarrega una variable per un any concret
}
\details{
Funció dissenyada per ús intern a la funció \code{donwload_var()}.
}
