% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mde.R
\name{get_mde}
\alias{get_mde}
\title{Obté un model digital d'elevacions per l'àrea d'estudi}
\usage{
get_mde(area_estudi, path, zoom = 12, ...)
}
\arguments{
\item{area_estudi}{Arxiu raster amb l'àrea d'estudi}

\item{path}{Destí on es vol guardar el raster}

\item{zoom}{La resolució del raster. \code{?elevatr::get_elev_raster} per més info.}

\item{...}{altres arguments per \code{elevatr::get_elev_raster}.}
}
\value{
El destí de l'arxiu de forma invisible
}
\description{
Obté un model digital d'elevacions per l'àrea d'estudi
}
