% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mde.R
\name{get_terrain}
\alias{get_terrain}
\title{Obté variables derivades del model digital d'elevacions}
\usage{
get_terrain(
  mde,
  var = c("slope", "aspect", "TPI", "TRI", "TRIriley", "TRIrmsd", "roughness", "flowdir"),
  path
)
}
\arguments{
\item{mde}{Arxiu ràster amb un model digital d'elevacions}

\item{var}{Variable a calcular. Una o més d'aquestes opcions: slope, aspect, TPI, TRI, TRIriley, TRIrmsd, roughness, flowdir. Consulta \link[terra]{terrain} per més info.}

\item{path}{Destí on es vol guardar el ràster resultant.}
}
\value{
El destí de l'arxiu de forma invisible.
}
\description{
Obté variables derivades del model digital d'elevacions
}
