% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpola dades de les estacions meteorològiques}
\usage{
interpolate(
  stations_data,
  stations_features,
  grid_features,
  allow_negative = TRUE,
  smooth = TRUE,
  area_estudi,
  method = c("lm", "lasso")
)
}
\arguments{
\item{stations_data}{Dades climàtiques de les estacions, obtingudes amb \link{download_var}}

\item{stations_features}{Covariables de les estacions, obtingudes amb \link{calculate_features}}

\item{grid_features}{Covariables de la malla de punts, obtingudes amb \link{calculate_features}}

\item{allow_negative}{Si és FALSE, els valors negatius es converteixen a zero.}

\item{smooth}{Si és TRUE, es fa un smooth dels resultats utilitzant una mitjana mòbil de 3x3}

\item{area_estudi}{Arxiu ràster amb l'àrea d'estudi. Només necessari si \code{smooth = TRUE}.}

\item{method}{Opció per utilitzar regressió linial o regressió lasso en la interpolació.}
}
\value{
Una tidytable amb els valors de la variable interpolada a cada punt de la malla
}
\description{
Interpola dades de les estacions meteorològiques
}
