% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_grid.R
\name{rasterize_points}
\alias{rasterize_points}
\title{Converteix la malla de punts a raster}
\usage{
rasterize_points(punts_sf, field, template, df = NULL)
}
\arguments{
\item{punts_sf}{Objecte \code{sf} que conté la malla de punts, si \code{df = NULL}, \code{punts_sf} ha de contenir \code{field}.}

\item{field}{Caràcter amb el nom del camp a rasteritzar de \code{punts_sf} o de \code{df}.}

\item{template}{Arxiu ràster per utilitzar com a plantilla.}

\item{df}{\code{data.frame} que conté \code{field} si aquest no es troba a \code{punts_sf}.}
}
\value{
Objecte \code{stars}.
}
\description{
Converteix la malla de punts a raster
}
